Add CuGraphEdgeMap -> CuGraphEdgeSet translator and translation test
This commit adds translators and tests for CuGraphEdgeMap -> CuGraphEdgeSet.
